Сделал рефакторинг кода
