+++ index.html
@@ -62,7 +62,7 @@
 
                   <!-- Изменение глубины эффекта, накладываемого на изображение -->
                   <fieldset class="img-upload__effect-level  effect-level">
-                    <input class="effect-level__value" type="number" step="any" name="effect-level" value="">
+                    <input class="effect-level__value" type="number" step="any" name="effect-level" value="5">
                     <div class="effect-level__slider"></div>
                   </fieldset>
