+++ js/pictures-popup.js
@@ -63,7 +63,7 @@ bigPictureClose.addEventListener('click', () => {
   commentsCountShown = 0;
 });
 
-window.addEventListener('keydown', (evt) => {
+document.addEventListener('keydown', (evt) => {
   if (evt.key === 'Escape') {
     bigPicture.classList.add('hidden');
     body.classList.remove('.modal-open');
